1. Porting Elastos.Runtime, Elastos.Libcore and Elastos.Framework.Core to Android O; 2. Calculator app run succeeded on google pixel phone with Android O

+++ Sources/Elastos/Frameworks/Droid/Base/Core/src/elastos/droid/internal/content/CNativeLibraryHelper.cpp
@@ -760,8 +760,8 @@ static install_status_t copyFileIfChanged(void* arg, ZipFileRO* zipFile, ZipEntr
     String nativeLibPath((char*) arg);
 
     size_t uncompLen;
-    long when;
-    long crc;
+    uint32_t when;
+    uint32_t crc;
     time_t modTime;
 
     if (!zipFile->getEntryInfo(zipEntry, NULL, &uncompLen, NULL, NULL, &when, &crc)) {
